закрыл баг

+++ BrowserInteractLabeler/Infrastructure/Buffers/CacheAnnotation.cs
@@ -1,5 +1,7 @@
 using BrowserInteractLabeler.Common;
 using BrowserInteractLabeler.Common.DTO;
+using Serilog;
+using ILogger = Serilog.ILogger;
 
 namespace BrowserInteractLabeler.Infrastructure;
 
@@ -9,20 +11,20 @@ public class CacheAnnotation
     private Annotation _lastAnnotation = new();
     private List<Annotation> _annotations = new();
     private int _lastIdDb = -1;
-    public CacheAnnotation(    IRepository repository)
+    private readonly ILogger _logger = Log.ForContext<CacheAnnotation>();
+
+    public CacheAnnotation(IRepository repository)
     {
         _repository = repository ?? throw new ArgumentNullException(nameof(repository));
-       
     }
-    
-   public async Task<(bool result,Annotation annot)> GetEditAnnotation()
+
+    public async Task<(bool result, Annotation annot)> GetEditAnnotation()
     {
-     
-        var annot = _annotations.FirstOrDefault(p => p.State!=StateAnnot.Finalized);
+        var annot = _annotations.FirstOrDefault(p => p.State != StateAnnot.Finalized);
         if (annot is not null)
-            return (true,annot);
-        
-        return (false,new Annotation());
+            return (true, annot);
+
+        return (false, new Annotation());
     }
 
     public async Task UpdateAnnotation(Annotation annotation)
@@ -38,54 +40,68 @@ public class CacheAnnotation
     }
 
 
-    public  Annotation[] GetAllAnnotations(int imagesId)
+    public Annotation[] GetAllAnnotations(int imagesId)
     {
-        return _annotations.Where(p=>p.ImageFrameId==imagesId).ToArray();
+        return _annotations.Where(p => p.ImageFrameId == imagesId).ToArray();
     }
 
     public async Task SaveAnnotationsOnSqlAsync(int imagesId)
     {
-        var removeAnnot =  await _repository.GetAnnotationsFromImgIdAsync(imagesId);
+        var removeAnnot = await _repository.GetAnnotationsFromImgIdAsync(imagesId);
+
+        var equalAnnotation = removeAnnot.Equality(_annotations.ToArray());
+        if (equalAnnotation)
+            return;
+        
+
+        _logger.Debug("[SaveAnnotationsOnSqlAsync] " +
+                      "Save annotations in Img:{imagesId} count annotations:{CountAnnotations}", imagesId,
+            _annotations.Count);
+
+
         await _repository.DeleteAnnotationsAsync(removeAnnot);
         foreach (var annotation in _annotations)
         {
             annotation.State = StateAnnot.Finalized;
+            annotation.Id = 0;
         }
-   
+
         await _repository.SaveAnnotationsAsync(_annotations.ToArray());
+
+        var allAnnot = await _repository.GetAnnotationsFromImgIdAsync(imagesId);
+        _annotations = allAnnot.CloneDeep().ToList();
     }
 
+ 
     public void RemoveLastAnnotation(int imagesId)
     {
-        var last = _annotations.LastOrDefault(p => p.ImageFrameId==imagesId);
+        var last = _annotations.LastOrDefault(p => p.ImageFrameId == imagesId);
         if (last is null)
             return;
         _lastAnnotation = last;
         _annotations.Remove(_lastAnnotation);
-
     }
 
     public void RestoreLastAnnotation(int imagesId)
     {
-        if(_lastAnnotation.Id<0 || _lastAnnotation.ImageFrameId !=imagesId )
+        if (_lastAnnotation.Id < 0 || _lastAnnotation.ImageFrameId != imagesId)
             return;
-        
+
         _annotations.Add(_lastAnnotation);
         _lastAnnotation = new Annotation();
     }
 
     public async Task LoadAnnotationsSlowStorageAsync(int imagesId)
     {
-        var allAnnot =  await _repository.GetAnnotationsFromImgIdAsync(imagesId);
-        _annotations =allAnnot.ToList();
-    
+        var allAnnot = await _repository.GetAnnotationsFromImgIdAsync(imagesId);
+        _annotations = allAnnot.CloneDeep().ToList();
+
         _lastIdDb = await _repository.GetLastIndexAnnotation();
- 
     }
 
     public void DeleteAnnotation()
     {
-        var last = _annotations.LastOrDefault(p => p.State!=StateAnnot.Finalized);
+        var last = _annotations.LastOrDefault(p => p.State != StateAnnot.Finalized);
         if (last is null)
             return;
         _lastAnnotation = last;
@@ -95,7 +111,7 @@ public class CacheAnnotation
     private void CreateNewAnnot(int imagesId)
     {
         _lastIdDb += 1;
-        var annot =new Annotation()
+        var annot = new Annotation()
         {
             Id = _lastIdDb,
             Points = new List<PointF>(),
@@ -105,20 +121,28 @@ public class CacheAnnotation
         };
         _annotations.Add(annot);
     }
-    
-    public  void EventEditAnnotForceCreateNew(int imagesId)
+
+    /// <summary>
+    ///     key q,w,a,s
+    /// </summary>
+    /// <param name="imagesId"></param>
+    public void EventEditAnnotForceCreateNew(int imagesId)
     {
         foreach (var annotation in _annotations)
         {
-            annotation.State =StateAnnot.Finalized;
+            annotation.State = StateAnnot.Finalized;
         }
-        
-        CreateNewAnnot(imagesId);
 
+        CreateNewAnnot(imagesId);
     }
-    public  void EventEditAnnot(int imagesId)
+
+    /// <summary>
+    ///     key [e]
+    /// </summary>
+    /// <param name="imagesId"></param>
+    public void EventEditAnnot(int imagesId)
     {
-        var last = _annotations.LastOrDefault(p => p.State!=StateAnnot.Finalized);
+        var last = _annotations.LastOrDefault(p => p.State != StateAnnot.Finalized);
         if (last is not null)
         {
             last.State = StateAnnot.Finalized;
@@ -126,13 +150,12 @@ public class CacheAnnotation
         }
 
         CreateNewAnnot(imagesId);
-
     }
 
     public bool SetActiveAnnot(int idAnnot)
     {
-        var current = _annotations.LastOrDefault(p => p.Id==idAnnot);
-     
+        var current = _annotations.LastOrDefault(p => p.Id == idAnnot);
+
 
         if (current is not null)
         {
@@ -140,6 +163,7 @@ public class CacheAnnotation
             {
                 annotation.State = StateAnnot.Finalized;
             }
+
             current.State = StateAnnot.Active;
             return true;
         }
@@ -149,7 +173,7 @@ public class CacheAnnotation
 
     public void SetActiveIdLabel(int id)
     {
-        var current = _annotations.LastOrDefault(p => p.State!=StateAnnot.Finalized);
+        var current = _annotations.LastOrDefault(p => p.State != StateAnnot.Finalized);
         if (current is not null)
         {
             current.LabelId = id;
